net: lib: download_client: move header under /net
Move header file under /net, since it is a networking library.
